backend/services/db_cleanup/requirements.txt
flask==3.0.0
flask-cors==4.0.0
psycopg2-binary==2.9.9
APScheduler==3.10.4
python-dotenv==1.0.0

# Testing dependencies
pytest==7.4.3
pytest-cov==4.1.0
pytest-mock==3.12.0
